requirements.txt
pandas==2.0.3
numpy==1.24.4
scikit-learn==1.3.0
joblib==1.3.1
pyarrow==13.0.0
fastparquet==2023.7.0
requests==2.31.0
beautifulsoup4==4.12.2
boto3==1.28.38
python-dotenv==1.0.0
streamlit==1.25.0
plotly==5.16.1
matplotlib==3.7.2
networkx==3.1
tensorflow==2.13.0
pytorch==2.0.1
Scheduling
schedule==1.2.0
Testing
pytest==7.4.0